requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
defusedxml==0.7.1
Django==5.1.4
django-allauth==0.41.0
django-modeltranslation==0.19.11
idna==3.10
oauthlib==3.2.2
pillow==11.0.0
python3-openid==3.2.0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
